app/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
prometheus-client==0.19.0
psutil==5.9.6
requests==2.31.0
python-dotenv==1.0.0
